+++ src/answer.js
@@ -21,16 +21,6 @@ add('norwegen', 'Darum sind die auch nicht in der EU, weil die am Leben vorbeila
 add('zu hause', 'Sollen wir nach Hause fahr\'n oder wat?');
 add('kann nichts', 'Junge, Junge, Junge, Junge, Junge, Junge, Junge, Junge, Junge, Junge!');
 add('kokowei', require('./kokoBuilder.js').getKoko);
-add(['a5','rhein','main','breakdance','club','feiern'],
-  function () {
-    var message = {
-      msgString : 'Also safe?',
-      params : {
-        as_user : false,
-        icon_emoji : ':derfickerx:',
-        username : 'Chirs'
-    }};
-    return message;
-});
+add(['a5','rhein','main','breakdance','club','feiern'],require('./chris.js').safe);
 
 module.exports = dict;
